add exceptions

+++ thirdlab-standalone/src/main/java/ru/itmo/webservices/thirdlab/standalone/ArticleWebService.java
@@ -1,6 +1,7 @@
 package ru.itmo.webservices.thirdlab.standalone;
 
 import ru.itmo.webservices.thirdlab.standalone.dao.ArticlesDao;
+import ru.itmo.webservices.thirdlab.standalone.exceptions.*;
 import ru.itmo.webservices.thirdlab.standalone.pojo.Article;
 
 import javax.jws.WebMethod;
@@ -25,7 +26,35 @@ public class ArticleWebService {
                               @WebParam(name = "hIndex") Long hIndex,
                               @WebParam(name = "articleName") String articleName,
                               @WebParam(name = "articleDesc") String articleDesc,
-                              @WebParam(name = "dateAdded") Long dateAdded) {
+                              @WebParam(name = "dateAdded") Long dateAdded) throws InsertingException, InvalidCreatingParametersException {
+        ArticleServiceFault fault = ArticleServiceFault.defaultInstance();
+        fault.setMessage("Invalid creating parameter");
+
+        if (authorId == null || authorId.trim().isEmpty()) {
+            fault.setMessage("Parameter authorId cannot be null or empty");
+            throw new InvalidCreatingParametersException("Invalid creating parameter", fault);
+        }
+
+        if (hIndex == null) {
+            fault.setMessage("Parameter hIndex cannot be null");
+            throw new InvalidCreatingParametersException("Invalid creating parameter", fault);
+        }
+
+        if (articleName == null || articleName.trim().isEmpty()) {
+            fault.setMessage("Parameter articleName cannot be null or empty");
+            throw new InvalidCreatingParametersException("Invalid creating parameter", fault);
+        }
+
+        if (articleDesc == null || articleDesc.trim().isEmpty()) {
+            fault.setMessage("Parameter articleDesc cannot be null or empty");
+            throw new InvalidCreatingParametersException("Invalid creating parameter", fault);
+        }
+
+        if (dateAdded == null) {
+            fault.setMessage("Parameter dateAdded cannot be null");
+            throw new InvalidCreatingParametersException("Invalid creating parameter", fault);
+        }
+
         ArticlesDao dao = new ArticlesDao();
         return dao.insert(authorId, hIndex, articleName, articleDesc, dateAdded);
     }
@@ -36,14 +65,24 @@ public class ArticleWebService {
                               @WebParam(name = "hIndex") Long hIndex,
                               @WebParam(name = "articleName") String articleName,
                               @WebParam(name = "articleDesc") String articleDesc,
-                              @WebParam(name = "dateAdded") Long dateAdded) {
+                              @WebParam(name = "dateAdded") Long dateAdded) throws IllegalIdException, InvalidEntityException {
         ArticlesDao dao = new ArticlesDao();
+        if (id == 0) {
+            ArticleServiceFault fault = ArticleServiceFault.defaultInstance();
+            fault.setMessage("Parameter id cannot be null or empty");
+            throw new IllegalIdException("Parameter id cannot be null or empty", fault);
+        }
         return dao.update(id, authorId, hIndex, articleName, articleDesc, dateAdded);
     }
 
     @WebMethod
-    public int deleteArticle(@WebParam(name = "id") int id) {
+    public int deleteArticle(@WebParam(name = "id") int id) throws IllegalIdException, InvalidEntityException {
         ArticlesDao dao = new ArticlesDao();
+        if (id == 0) {
+            ArticleServiceFault fault = ArticleServiceFault.defaultInstance();
+            fault.setMessage("Parameter id cannot be null or empty");
+            throw new IllegalIdException("Parameter id cannot be null or empty", fault);
+        }
         return dao.delete(id);
     }
 }
